Bug fix in scene addRootNode for non-dirty scene nodes.
Previously adding a root node which had been removed from the scene
and added back again without making any modifications meant that
with no dirty flags set for extents the node would not be added
to the spatial map.

Setting worldExtentsUpdate to true ensures that on scene update,
the node will be added to the spatial maps regardless of having
clean or dirty extents.